[IOSA-23] Biometric authentication + a lot of refactoring
