+++ SDK Showcase/SDK Showcase/Helpers/SDKError.swift
@@ -6,6 +6,7 @@ enum SDKError: Int, Error {
     case unknown
     case accountDeregistered = 9003
     case registrationCancelled = 9006
+    case biometricAuthenticationFallback = 9030
     case statelessDisabled = 9034
     
     init(_ error: Error) {
@@ -14,6 +15,8 @@ enum SDKError: Int, Error {
             self = .accountDeregistered
         case 9006:
             self = .registrationCancelled
+        case 9030:
+            self = .biometricAuthenticationFallback
         case 9034:
             self = .statelessDisabled
         default:
