+++ SDK Showcase/SDK Showcase/Views/AuthenticatorsSheet.swift
@@ -7,24 +7,6 @@ struct AuthenticatorsSheet: View {
         @Injected var appState: AppState
         return appState.system
     }()
-    var sdkInteractor: SDKInteractor {
-        @Injected var interactors: Interactors
-        return interactors.sdkInteractor
-    }
-    var authRegistrationInteractor: AuthenticatorRegistrationInteractor {
-        @Injected var interactors: Interactors
-        return interactors.authenticatorRegistrationInteractor
-    }
-    private var sheetTitle: String {
-        if let _ = selectedOption?.name {
-            return "Select the authenticator for authentication"
-        } else if let authenticatedId = system.userState.userId, authRegistrationInteractor.unregisteredAuthenticatorNames(for: authenticatedId).count > 0 {
-            return "Select the authenticator to register"
-        } else {
-            return "No authenticators left to register"
-        }
-    }
-    
     var showConfirmationDialog: Binding<Bool>
     var selectedOption: Selection?
     var body: some View {
@@ -34,8 +16,8 @@ struct AuthenticatorsSheet: View {
             if let profileId = selectedOption?.name  {
                 ForEach(authRegistrationInteractor.authenticatorNames(for: profileId), id: \.self) { name in
                     Button(name) {
-                        sdkInteractor.authenticateUser(profileName: profileId,
-                                                       optionName: name)
+                        authInteractor.authenticateUser(profileName: profileId,
+                                                        optionName: name)
                     }
                 }
             } else if let authenticatedId = system.userState.userId {
@@ -45,6 +27,26 @@ struct AuthenticatorsSheet: View {
                     }
                 }
             }
+        }
+    }
+}
+
+private extension AuthenticatorsSheet {
+    var authRegistrationInteractor: AuthenticatorRegistrationInteractor {
+        @Injected var interactors: Interactors
+        return interactors.authenticatorRegistrationInteractor
+    }
+    var authInteractor: AuthenticatorInteractor {
+        @Injected var interactors: Interactors
+        return interactors.authenticatorInteractor
+    }
+    var sheetTitle: String {
+        if let _ = selectedOption?.name {
+            return "Select the authenticator for authentication"
+        } else if let authenticatedId = system.userState.userId, authRegistrationInteractor.unregisteredAuthenticatorNames(for: authenticatedId).count > 0 {
+            return "Select the authenticator to register"
+        } else {
+            return "No authenticators left to register"
         }
     }
 }
